requirements.txt
# vLLM-Ascend Python Dependencies

# Core dependencies
torch==2.1.0
torch_npu==2.1.0.post3

# vLLM requirements
ray==2.9.0
transformers==4.40.0
tokenizers==0.19.1
sentencepiece==0.2.0

# API server
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3

# Utilities
numpy==1.24.3
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.3

# Monitoring
prometheus-client==0.20.0

# Testing (optional, can be moved to requirements-dev.txt)
pytest==8.1.1
pytest-asyncio==0.23.5

# Hugging Face
huggingface-hub==0.21.4
